requirements.txt
aniso8601==7.0.0
asgiref==3.5.1
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.4
django-cors-headers==3.12.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
Faker==12.0.1
Flask==2.1.1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.3
importlib-metadata==4.11.3
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.1
Markdown==3.3.7
MarkupSafe==2.1.1
mixer==7.2.2
packaging==21.3
promise==2.3
psycopg2==2.8.6 
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
